add default column
